update objectdetection/for_atlas200dk_1.7x.0.0/objectdetection/src/object_detect.cpp.
Former-commit-id: 69fb3e3d4f77f0bcc3661050867b3f536fd0f5fa

+++ objectdetection/for_atlas200dk_1.7x.0.0/objectdetection/src/object_detect.cpp
@@ -251,7 +251,7 @@ Result ObjectDetect::Postprocess(ImageData& image, aclmdlDataset* modelOutput,
     float* detectData = (float *)GetInferenceOutputItem(dataSize, modelOutput,
                                                         kBBoxDataBufId);
 
-    float* boxNum = (float *)GetInferenceOutputItem(dataSize, modelOutput,
+    uint32_t* boxNum = (uint32_t *)GetInferenceOutputItem(dataSize, modelOutput,
                                                     kBoxNumDataBufId);
     if (boxNum == nullptr) return FAILED;
